Read source code from a file, a REPL, and handle errors
If the interpreter is given a second argument, it treats that second
argument as a file, from which to read source code from. Otherwise, it
brings up a REPL where one can type code in. Currently it only
tokenizes the source code and prints out the tokens, but it's progress!